Client implementation for deleting messages
This is basically only a reference for now, since the client is in the
process of being rewritten. Also, this does not follow the spec as
described in #66 (i.e. no "message action" dropdown); rather, it
replaces the error message that shows up when a message is edited and a
zero-length input is entered.