Add 'routeready' event
This is emitted *after* app.js does everything it needs to do with the
route event. It's super handy and fixes like a billion (not filed) bugs.

Without this, we have to choose a "sensible default" for whether the
server is private or not, and whether it runs on HTTPS or not. But that
default isn't actually right - you end up having mixed content, or
errors related to fetching an API without passing a session ID, or even
just not having messages fetched at all, etc. This gets rid of those by
giving a different event - routeready - for those actions to wait on.

PS, routeready also waits for the WebSocket to be ready, since pretty
much everything doesn't function correctly/completely without it (e.g.
old messages showing up, but new messages not, because the WS isn't
connected and so isn't reveing "received a message" events).