Make tests run concurrently better
Unfortunately, openPort doesn't work well when we're dealing with a lot
of things running concurrently (and in separate processes
simultaneously). Instead, we just use a base port number constant in
each test file and increase that for each spawned server. (So, 21000 in
tests/auth.js, 22000 in tests/middleware.js, 23000 in the next test file
we make, etc -- these need to be very far from each other because we
don't want to have conflicts and, as I mentioned, openPort doesn't work.)