fromCLI arg for addRole to fix prio-ing admin role
Before this, the make-admin command wouldn't work unless the admin role
already existed (and was present in the role prioritization list).

addRole wouldn't work without an actor user ID provided because it
positions the new role as prioritized beneath the passed user's highest
role. This makes sense in most cases, but from the CLI, there *is* no
actor at all - so instead, we now place the role at the very top, which
is where the admin role should be.